Is this pinning param correct?
Quickly looking in the given article I don't see why this p should be multiplied by 2*PI. Also maybe would be better to stick with only the "dp" pinning parameter for consistence? Which I think should have the units (1/m) and not (rad/m) as written in SingleLayer class. Also maybe change Ku to Ks in the Material calss?