GetFnn and threshold field parameters correction
I added GetFnn function to get the Fnn parameter from dispersion so I could use it for the Ak and Bk parameters which are the used in coupling parameter for the threshold field of parallel pumping. The definitions of Ak and Bk were adjusted to utilize the Fnn parameter and the threshold field calculation was rewritten for better readibility. The calculated threshold curve now holds with what is in Bjorn Heinz dissertation