Update pyproject.toml after first upload to PyPI
After 1h of fight, I managed to upload v1.0.0 to PyPI (as it is the current master branch), but I transferred the changes here to the latest branch.

To make it work, I needed these changes together with downgraded twine to 6.0.1 instead of 6.1.0. Otherwise I followed the instructions in https://packaging.python.org/en/latest/tutorials/packaging-projects/