up SL
Changes definition of phi_H to be relative to kxi (remember kxi || x). I think it was incorrectly used as we also need to take into account magnetization phi as an angle from kxi, not angle of kxi from M -> kxi is the reference!

I also added the methods requested in #34