up BLS example, fix SL to support sweeps
Before it was quite nice that some parameters could be supplied in the same shape as kxi to dramatically speed up calculations (in SL only unfortunately), but after the OOP fix, this was lost. So here is my attempt on a fix. The sweep calculations still need to be tested though, because I don't trust it that much. (copilot helped me a lot)